+++ lib/features/review_page/data/repositories/elaboration/elaboration_repository_impl.dart
@@ -3,9 +3,9 @@ import 'package:dartz/dartz.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
 import 'package:u_do_note/core/error/failures.dart';
-import 'package:u_do_note/features/review_page/data/datasources/elaboration_remote_datasource.dart';
+import 'package:u_do_note/features/review_page/data/datasources/elaboration/elaboration_remote_datasource.dart';
 import 'package:u_do_note/features/review_page/data/models/elaboration.dart';
-import 'package:u_do_note/features/review_page/domain/repositories/elaboration_repository.dart';
+import 'package:u_do_note/features/review_page/domain/repositories/elaboration/elaboration_repository.dart';
 
 class ElaborationImpl implements ElaborationRepository {
   final ElaborationRemoteDataSource _elaborationRemoteDataSource;
