+++ lib/features/review_page/data/repositories/feynman/feynman_technique_repository_impl.dart
@@ -2,9 +2,10 @@ import 'package:dart_openai/dart_openai.dart';
 import 'package:dartz/dartz.dart';
 
 import 'package:u_do_note/core/error/failures.dart';
-import 'package:u_do_note/features/review_page/data/datasources/feynman_remote_datasource.dart';
+import 'package:u_do_note/features/review_page/data/datasources/feynman/feynman_remote_datasource.dart';
 import 'package:u_do_note/features/review_page/data/models/feynman.dart';
-import 'package:u_do_note/features/review_page/domain/repositories/feynman_technique_repository.dart';
+
+import 'package:u_do_note/features/review_page/domain/repositories/feynman/feynman_technique_repository.dart';
 
 class FeynmanTechniqueImpl implements FeynmanTechniqueRepository {
   final FeynmanRemoteDataSource _feynmanRemoteDataSource;
