+++ lib/features/review_page/data/repositories/pomodoro/pomodoro_technique_repository_impl.dart
@@ -3,7 +3,7 @@ import 'package:dartz/dartz.dart';
 import 'package:u_do_note/core/error/failures.dart';
 import 'package:u_do_note/features/review_page/data/datasources/pomodoro/pomdoro_remote_datasource.dart';
 import 'package:u_do_note/features/review_page/data/models/pomodoro.dart';
-import 'package:u_do_note/features/review_page/domain/repositories/pomodoro_technique_repository.dart';
+import 'package:u_do_note/features/review_page/domain/repositories/pomodoro/pomodoro_technique_repository.dart';
 
 class PomodoroTechniqueRepositoryImpl implements PomodoroTechniqueRepository {
   final PomodoroRemoteDataSource _pomodoroRemoteDataSource;
