+++ lib/features/review_page/domain/usecases/leitner/generate_flashcards.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:u_do_note/core/error/failures.dart';
 import 'package:u_do_note/features/review_page/data/models/leitner.dart';
-import 'package:u_do_note/features/review_page/domain/repositories/leitner_system_repository.dart';
+import 'package:u_do_note/features/review_page/domain/repositories/leitner/leitner_system_repository.dart';
 
 class GenerateFlashcards {
   final LeitnerSystemRepository _leitnerSystemRepository;
