+++ lib/features/review_page/domain/usecases/leitner/get_old_flashcards.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:u_do_note/core/error/failures.dart';
 import 'package:u_do_note/features/review_page/data/models/leitner.dart';
-import 'package:u_do_note/features/review_page/domain/repositories/leitner_system_repository.dart';
+import 'package:u_do_note/features/review_page/domain/repositories/leitner/leitner_system_repository.dart';
 
 class GetOldFlashcards {
   final LeitnerSystemRepository _repository;
